Add notification grace period after carbon message
Add the option to suspend vibration/ringtone alerts for a short amount
of time after we receive a carbon message. The idea here is that we
don't want our phone to be buzzing in our phone while we're actively
chatting from a different client.

The option is enabled by default, grace period is set to 60s for now.
This may be a target for future fine tuning.