Added tiled MUC icons
Now, MUCs have icons with up to 4 tiles representing MUC members,
similar to the gmail app.

Some caveats:
As of now, no ordering is imposed on the members, we simply pick the
first (up to) 4 returned by the MucOptions object. This could be done
better, e.g. by picking more important members first (based on
affil/role), or even going so far as to order them by how recently they
sent a message in the MUC.
Also, the code generating the tiles is really messy right now (tons of
copy and paste). It seems to work though. I will clean this up soon.