incease undo timeout
I like the swiping out feature, but for me the display time of undo swiping out a conversation is too short. 
Remember, this behaviour is a new feature, and a lot of users doesn't read changelogs, they only want to use the app. I have some doubts especially for MUCs. Remember if somebody uses this function unconsciously she/he won't know whats happening and until she/he could realize what this swiping out does, the toast disappears and the conference was left. Then the user is angry because she/he has left the conference, without any notice. Perhaps it would be better to confirm swiping out with the hint of receiving no messages after doing this? I think a global increasing of the display time is the easiest was to give the user a chance to undo that, especially for conferences where the room was permanently left with the consequence of not getting any messages, different to single conversations, where you still receive messages.

Perhaps we can think about swiping out only to the left, because there is a swiping action between chat and chatlist over the right border. 

What are you thinking about?