Use a RecyclerView for list of gateway options
If there are no gateways in the current account's roster, just show the
old "Jabber ID" label.  Otherwise show a list of toggle buttons.  These
buttons show the type of their gateway and change the JID input to
not autocomplete and to show a hint matching the jabber:id:gateway
prompt.