Adding handsfreeservice
Functionality:
The HandsFree service provides a speakMessage service, which subsiquently can offer the user the possibility to reply using speech recognition.

The service will respect the users interuptions settings (e.g. Do Not Desturb)
It will also mute and/or stop vice recognition on phone calls
    (Both of these services need permissions)
Finally it tries to 'play nicely' with other audio users on the device.