Update strings.xml
Make clear to the user that the file will only be delete locally. The current message "Delete image", could led the user to believe that the file does not reside any longer on the server, thus will be removed not only locally but from the server as well. 
This change will still not allow the user to request for files to be removed form the Server, thus he would need to contact the System Administrator in order to get his files removed.
Moreover, this changes needs to be done to the other languages. 
See https://github.com/siacs/Conversations/issues/3193#issuecomment-425697655 for further explanation.