Extend the OMEMO auto expiry to 3 months
This is the time after which Conversations will kick an OMEMO key from a
users key list.

Note that this is about all keys in the list, not just the one
Conversations is using. So this could be a desktop client that the user
only uses on certain machines that will start losing history after 14
days.