strings: Simplify minutes_ago string
Instead of "15 minutes ago", just say "15 m", which is considerably
shorter without adding ambiguity or confusion (well the message isn't 15
meters away...). The main benefit, aside from concision alone, is that
these date stamps are metadata, so by shrinking and simplifying we
reduce visual clutter and focus the user on the message content.

Signal does this as well. See also
https://github.com/siacs/Conversations/issues/3544

The English and Spanish strings are updated, since those are the
languages I know and am comfortable that the idiom makes sense in. Other
languages are not touched as I don't know in which languages this
simplification makes sense, but by retaining the minutes_ago string,
phones in languages other than English and Spanish will retain the
current behaviour.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>