strings: Drop "with OMEMO" in the input box
Given that OTR is no longer support and OpenPGP is for advanced users
with special needs only, it makes sense to normalize the idea that
"encrypted = OMEMO" within the context of XMPP. But at that point,
there's no need to highlight OMEMO itself -- that becomes an
implementation detail for the user. So it makes sense to simplify the
input box text to simply "Send encrypted message" when OMEMO is active,
dropping the "with OMEMO" piece which adds clutter and confusion for
users who aren't familiar with the protocol. (which ideally would be
most of them! No need to be a cryptographer to use crypto.)

String changed in English and Romance languages (where I could do so
with confidence that I wasn't breaking things gramatically). Other
languages retain the current behvaiour, but I would encourage
translators to adopt this as well.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>